apps/bridge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "printfarm-bridge"
version = "0.1.0"
description = "PrintFarm Bridge — Bambu Lab LAN MQTT agent"
requires-python = ">=3.9"
dependencies = [
    "paho-mqtt>=2.0.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
camera = [
    "opencv-python-headless>=4.9.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "aiohttp>=3.9.0",
]

[project.scripts]
printfarm-bridge = "src.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
source = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
